feat : @getter, @allargsconstructor를 이용하여 도메인 메소드 추상화
