+++ src/main/java/model/UserPrincipal.java
@@ -1,26 +1,19 @@
 package model;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
 public class UserPrincipal {
 
     public String userId;
 
     public String name;
 
-    public UserPrincipal(String userId, String username) {
-        this.userId = userId;
-        this.name = username;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
 
     public void setUserId(String userId) {
         this.userId = userId;
-    }
-
-    public String getName() {
-        return name;
     }
 
     public void setName(String username) {
